Działa tworzenie i wypełnianie bazy. Dodano klase HibernateUtil generujaca service factory i konfiguracje
